code_writer/pyproject.toml
[project]
name = "code-writer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.24",
    "openai>=2.24.0",
    "openai-agents>=0.10.4",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
